Add files via upload
A simple Number to Word Converter in Python

This is a program that converts integer numbers between 0 and 999 into words. For example, if the number 647 is inputted, the program will return the string "six hundred forty-seven".

Requirements
The program meets the following requirements:

The only input parameter of the function is an integer number between 0 and 999.
It uses dictionaries to match each digit with the corresponding word for units, tens, and hundreds.
It returns a string with the inputted quantity in words.
How to Use the Program

Open the program file in your preferred development environment (e.g., a text editor or an IDE).

Run the program.
You will be prompted to enter an integer number between 0 and 999.
Enter the number and press Enter.
The program will display the inputted quantity in words.
Enjoy using the number to word converter!